src/requirements.txt
Flask==2.3.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
cryptography>=41.0.3